all the Long are now Integer

+++ src/main/java/com/example/punished/expensetracker/repository/ExpenseRepository.java
@@ -3,5 +3,5 @@ package com.example.punished.expensetracker.repository;
 import com.example.punished.expensetracker.entity.Expense;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ExpenseRepository extends JpaRepository<Expense, Long> {
+public interface ExpenseRepository extends JpaRepository<Expense, Integer> {
 }
